requirements.txt
Django==2.2
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.1
django-filebrowser-no-grappelli==3.8.0
django-jet==1.0.8
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
